v3/services/rag-svc/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1
numpy>=1.26,<2.0
